More generic implementation
